fix product, voucher, order

+++ product/repository/repository.go
@@ -28,18 +28,16 @@ func (repo *repository) GetProduct(req model.ProductSearch) ([]model.Product, er
 	ctx, cancel := timeout.NewCtxTimeout()
 	defer cancel()
 
-	querySearchProduct := `select p.id, p.store_id, p.category_id, p.size_id, p.color_id ,p.name,p.subtitle ,p.description ,p.unit_price ,p.status ,p.stock ,p.sku ,p.weight ,p.created_at ,p.updated_at from products p  join categories c on c.id = p.category_id join product_sizes s on s.id = p.size_id join product_colors co on co.id = p.color_id where c.name like '%' || $1 || '%' and s.name like '%' || $2 || '%' and co.name like '%' || $3 || '%'`
+	querySearchProduct := `select p.id, p.store_id, p.category_id, p.size_id, p.color_id ,p.name, p.brand, p.subtitle, p.description ,p.unit_price ,p.status ,p.stock ,p.sku ,p.weight ,p.created_at ,p.updated_at from products p  join categories c on c.id = p.category_id join product_sizes s on s.id = p.size_id join product_colors co on co.id = p.color_id where c.name like '%' || $1 || '%' and p.name like '%' || $2 || '%' and p.brand like '%' || $3 || '%'`
 
-	result, err := repo.db.QueryContext(ctx, querySearchProduct, req.Brand, req.Name, req.Category)
-	if err != nil {
-		return []model.Product{}, errors.New("error get product")
-	}
+	result, err := repo.db.QueryContext(ctx, querySearchProduct, req.Category, req.Name, req.Brand)
+	failerror.FailError(err, "error query")
 
 	var data = []model.Product{}
 
 	for result.Next() {
 		var temp model.Product
-		result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
+		result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Brand, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
 		data = append(data, temp)
 	}
 
@@ -55,9 +53,13 @@ func (repo *repository) ShowProduct(id int) (model.Product, error) {
 	result, err := repo.db.QueryContext(ctx, query, id)
 	failerror.FailError(err, "error prepare")
 
-	var temp model.Product
+	var temp = model.Product{}
 	for result.Next() {
 		result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
+	}
+
+	if temp.Id <= 0 {
+		return temp, errors.New("product not found")
 	}
 
 	return temp, nil
@@ -76,6 +78,7 @@ func (repo *repository) CreateProduct(req []model.ProductReq) ([]model.Product, 
 			SizeID:      v.SizeID,
 			ColorID:     v.ColorID,
 			Name:        v.Name,
+			Brand:       v.Brand,
 			Subtitle:    v.Subtitle,
 			Description: v.Description,
 			UnitPrice:   v.UnitPrice,
@@ -93,10 +96,10 @@ func (repo *repository) CreateProduct(req []model.ProductReq) ([]model.Product, 
 		return nil, errors.New("failed publisher")
 	}
 
-	time.Sleep(1 * time.Second)
+	time.Sleep(3 * time.Second)
 
 	var resultss []model.Product
-	query := `select  id, store_id, category_id, size_id, color_id , name, subtitle , description , unit_price , status , stock ,sku , weight , created_at , updated_at from products p where sku = $1`
+	query := `select  id, store_id, category_id, size_id, color_id , name, brand,subtitle , description , unit_price , status , stock ,sku , weight , created_at , updated_at from products p where sku = $1`
 
 	stmt, err := repo.db.PrepareContext(ctx, query)
 	failerror.FailError(err, "error prepare")
@@ -108,7 +111,7 @@ func (repo *repository) CreateProduct(req []model.ProductReq) ([]model.Product, 
 
 		var temp model.Product
 		for result.Next() {
-			result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
+			result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Brand, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
 		}
 		if temp.Id == 0 {
 			continue
@@ -117,13 +120,13 @@ func (repo *repository) CreateProduct(req []model.ProductReq) ([]model.Product, 
 	}
 
 	if resultss == nil {
-		return nil, errors.New("error create product")
+		return []model.Product{}, errors.New("error create product")
 	}
 
 	return resultss, nil
 }
 
-func (repo *repository) UpdateProduct(req model.ProductReq) (model.Product, error) {
+func (repo *repository) UpdateProduct(req model.ProductUpd) (model.Product, error) {
 	ctx, cancel := timeout.NewCtxTimeout()
 	defer cancel()
 
@@ -135,14 +138,14 @@ func (repo *repository) UpdateProduct(req model.ProductReq) (model.Product, erro
 	log.Println(req)
 	time.Sleep(1 * time.Second)
 
-	query := `select * from products p where id = $1`
+	query := `select id, store_id, category_id, size_id, color_id , name, brand,subtitle , description , unit_price , status , stock ,sku , weight , created_at , updated_at from products p where id = $1`
 
 	result, err := repo.db.QueryContext(ctx, query, req.Id)
 	failerror.FailError(err, "error prepare")
 
 	var temp model.Product
 	for result.Next() {
-		result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
+		result.Scan(&temp.Id, &temp.StoreID, &temp.CategoryID, &temp.SizeID, &temp.ColorID, &temp.Name, &temp.Brand, &temp.Subtitle, &temp.Description, &temp.UnitPrice, &temp.Status, &temp.Stock, &temp.Sku, &temp.Weight, &temp.Created_at, &temp.Update_at)
 	}
 
 	if temp.Id == 0 {
@@ -152,23 +155,23 @@ func (repo *repository) UpdateProduct(req model.ProductReq) (model.Product, erro
 	return temp, nil
 }
 
-func (repo *repository) DeleteProduct(id int) error {
+func (repo *repository) DeleteProduct(id int) (int, error) {
 
 	ctx, cancel := timeout.NewCtxTimeout()
 	defer cancel()
 
-	// var idCheck int
-	// queryCheck := `select id from products where id = $1 returning id`
-	// err := repo.db.QueryRowContext(ctx, queryCheck, id).Scan(&idCheck)
-	// failerror.FailError(err, "error exec")
-
-	// if idCheck == 0 {
-	// 	return errors.New("id tidak ditemukan")
-	// }
-
-	query := `DELETE FROM products WHERE id = $1`
-	_, err := repo.db.ExecContext(ctx, query, id)
+	var idCheck int
+	queryCheck := `select id from products where id = $1`
+	err := repo.db.QueryRowContext(ctx, queryCheck, id).Scan(&idCheck)
 	failerror.FailError(err, "error exec")
 
-	return nil
+	if idCheck == 0 {
+		return 0, errors.New("product tidak ditemukan")
+	}
+
+	query := `DELETE FROM products WHERE id = $1`
+	_, err = repo.db.ExecContext(ctx, query, idCheck)
+	failerror.FailError(err, "error exec")
+
+	return idCheck, nil
 }
